requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.8.0
elevenlabs==0.2.26
exceptiongroup==1.1.3
executing==2.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.1
openai==1.1.1
opencv-python==4.8.1.78
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
requests==2.31.0
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.10
websockets==12.0
ttkbootstrap
pygame
python-dotenv
flask
